+++ tfe.go
@@ -166,6 +166,7 @@ type Client struct {
 	VariableSetVariables       VariableSetVariables
 	Workspaces                 Workspaces
 	WorkspaceRunTasks          WorkspaceRunTasks
+	Projects                   Projects
 
 	Meta Meta
 }
@@ -390,6 +391,7 @@ func NewClient(cfg *Config) (*Client, error) {
 	client.PolicySetParameters = &policySetParameters{client: client}
 	client.PolicySets = &policySets{client: client}
 	client.PolicySetVersions = &policySetVersions{client: client}
+	client.Projects = &projects{client: client}
 	client.RegistryModules = &registryModules{client: client}
 	client.RegistryProviderPlatforms = &registryProviderPlatforms{client: client}
 	client.RegistryProviders = &registryProviders{client: client}